requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
fastapi==0.109.2
flake8==7.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.11.1
pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
supabase==2.13.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.1
black==25.1.0
branca==0.8.1
cfgv==3.4.0
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.17.0
folium==0.19.5
httptools==0.6.4
identify==2.6.8
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy
orjson==3.10.15
pathspec==0.12.1
platformdirs==4.3.6
polyline==2.0.2
pre_commit==4.1.0
pydantic-extra-types==2.10.2
PyYAML==6.0.2
tomli==2.2.1
ujson==5.10.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.4
websockets>=11,<15
xyzservices==2025.1.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
deprecation==2.1.0
frozenlist==1.5.0
gotrue==2.11.4
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
multidict==6.1.0
postgrest==0.19.3
propcache==0.3.0
python-dateutil==2.9.0.post0
realtime==2.4.1
six==1.17.0
storage3==0.11.3
StrEnum==0.4.15
supafunc==0.9.3
yarl==1.18.3